examples/advanced/finance-compliance/requirements.txt
perseus-client
python-dotenv
neo4j
rdflib
